mm/phys: adopt SparseBitmap; platform-isolate FrameAllocator and Kpool
Rework the frame allocator and kernel page pool around
`SparseBitmap` and move their inner state into per-platform
modules under `hal/platform/{microvm,hyperlight}/`, so
`mm/phys/` has no cfg gates.

Allocator shape on microvm:

  * FrameAllocator is a thin facade owning a
    `hal::platform::frame::Inner` directly (single-owner — the
    upool consumes it during init and never shares the handle).
  * Kpool wraps `Rc<RefCell<hal::platform::kpool::Inner>>` so
    `KernelFrame::Drop` can safely dispatch a free through the
    shared RefCell (the guard prevents double-free during
    nested-drop scenarios).
  * Both `Inner`s use a single-chunk `SparseBitmap` at offset 0
    covering the dense identity / pool range. Foreign-address
    registration uses `SparseBitmap::set_or_add_chunk` to grow
    chunks on demand.
  * The BSS-backed storage that used to live as a `static mut`
    in `mm/phys/mod.rs` moves to `hal/platform/microvm/frame.rs`
    next to the code that owns it.

Allocator shape on hyperlight:

  * `hal/platform/hyperlight/{frame,kpool}.rs` exist as byte-
    identical mirrors of the microvm `Inner`s so the kernel
    compiles with `--features hyperlight`. Doc comments flag
    them as temporary scaffolding — the scratch-backed
    implementation lands with the rest of Nanvix-on-Hyperlight
    (CoW snapshot/restore); nothing in mm/phys/ will change at
    that point because the platform dispatch absorbs the
    difference.

Everything else is mechanical: tests still pass on microvm and
hyperlight, and the public types (`Kpool`, `KernelFrame`,
`FrameAllocator`) keep their shape for callers.
