+++ src/kernel/src/hal/platform/microvm/mod.rs
@@ -5,6 +5,8 @@
 // Modules
 //==================================================================================================
 
+pub mod frame;
+pub mod kpool;
 pub mod pvclock;
 
 //==================================================================================================
